move tools into inspect_ai.tool
note sure yet if we want to do this, perhaps they belong in inspect_ai.agent ?
