+++ examples/biology_qa.py
@@ -1,7 +1,8 @@
 from inspect_ai import Task, task
 from inspect_ai.dataset import FieldSpec, example_dataset
 from inspect_ai.scorer import model_graded_qa
-from inspect_ai.solver import generate, use_tools, web_search
+from inspect_ai.solver import generate, use_tools
+from inspect_ai.tool import web_search
 
 
 @task
