+++ src/inspect_ai/tool/__init__.py
@@ -11,8 +11,13 @@ from ._tool import Tool, ToolError, ToolResult, tool
 from ._tool_call import ToolCall
 from ._tool_choice import ToolChoice, ToolFunction
 from ._tool_info import ToolInfo, ToolParam
+from ._tools._execute import bash, python
+from ._tools._web_search import web_search
 
 __all__ = [
+    "bash",
+    "python",
+    "web_search",
     "tool",
     "Tool",
     "ToolError",
